BF: numpy 1.12 will be more strict on int indices
Numpy 1.12 pre-release insists in integer indices for slicing.

Along the way, there was a bug in the dmtx testing revealed by these
checks.

+++ nipy/algorithms/clustering/imm.py
@@ -234,7 +234,7 @@ class IMM(BGMM):
         if np.isscalar(kfold):
             aux = np.argsort(np.random.rand(n_samples))
             idx = - np.ones(n_samples).astype(np.int)
-            j = np.ceil(n_samples / kfold)
+            j = int(np.ceil(n_samples / kfold))
             kmax = kfold
             for k in range(kmax):
                 idx[aux[k * j:min(n_samples, j * (k + 1))]] = k
